Refactor _ei_improvement into gaussian_acquisition
